Add support for a desktop websocket server
