+++ app/src/main/java/com/emulnk/MainActivity.kt
@@ -24,16 +24,15 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.emulnk.core.DisplayHelper
-import com.emulnk.model.MatchConfidence
 import com.emulnk.core.OverlayConstants
 import com.emulnk.core.OverlayService
 import com.emulnk.core.UiConstants
 import com.emulnk.model.CustomOverlayConfig
 import com.emulnk.model.SavedOverlayConfig
-import com.emulnk.model.DisplayInfo
-import com.emulnk.model.SafeArea
+import com.emulnk.core.model.DisplayInfo
+import com.emulnk.core.model.SafeArea
 import com.emulnk.model.ScreenTarget
-import com.emulnk.model.SystemInfo
+import com.emulnk.core.model.SystemInfo
 import com.emulnk.model.ThemeConfig
 import com.emulnk.model.ThemeType
 import com.emulnk.model.resolvedType
