--- app/src/main/java/com/emulnk/model/ProfileConfig.kt
@@ -1,10 +1,0 @@
-package com.emulnk.model
-
-// Re-export core model types for backward compatibility with existing app code.
-// New code should import from com.emulnk.core.model directly.
-typealias ProfileConfig = com.emulnk.core.model.ProfileConfig
-typealias DataPoint = com.emulnk.core.model.DataPoint
-typealias MacroConfig = com.emulnk.core.model.MacroConfig
-typealias MacroStep = com.emulnk.core.model.MacroStep
-typealias BundleConfig = com.emulnk.core.model.BundleConfig
-typealias MatchConfidence = com.emulnk.core.model.MatchConfidence
