+++ app/src/main/java/com/emulnk/ui/viewmodel/MainViewModel.kt
@@ -3,41 +3,39 @@ package com.emulnk.ui.viewmodel
 import android.app.Application
 import android.content.Intent
 import android.util.Log
-import com.emulnk.BuildConfig
-import com.emulnk.EmuLnkApplication
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
+import com.emulnk.BuildConfig
+import com.emulnk.EmuLnkApplication
 import com.emulnk.core.DisplayHelper
 import com.emulnk.core.OverlayService
 import com.emulnk.core.SyncService
-import com.emulnk.core.MemoryConstants
 import com.emulnk.core.TelemetryConstants
 import com.emulnk.core.TelemetryService
+import com.emulnk.core.constants.MemoryConstants
+import com.emulnk.core.model.BatteryInfo
+import com.emulnk.core.model.GameData
+import com.emulnk.core.model.MatchConfidence
+import com.emulnk.core.model.SystemInfo
+import com.emulnk.core.model.ThermalInfo
 import com.emulnk.data.ConfigManager
-import com.emulnk.model.AppConfig
-import com.emulnk.model.BatteryInfo
-import com.emulnk.model.GameData
-import com.emulnk.model.OverlayBundle
 import com.emulnk.data.MigrationManager
+import com.emulnk.model.AppConfig
 import com.emulnk.model.CustomOverlayConfig
 import com.emulnk.model.GalleryIndex
 import com.emulnk.model.GalleryTheme
+import com.emulnk.model.OverlayBundle
 import com.emulnk.model.SavedOverlayConfig
 import com.emulnk.model.ScreenTarget
 import com.emulnk.model.StoreWidget
-import com.emulnk.model.WidgetConfig
-import com.emulnk.model.toWidgetConfig
-import com.emulnk.model.SystemInfo
-import com.emulnk.model.ThermalInfo
 import com.emulnk.model.ThemeConfig
 import com.emulnk.model.ThemeMeta
 import com.emulnk.model.ThemeType
 import com.emulnk.model.resolvedScreenTarget
 import com.emulnk.model.resolvedType
-import com.emulnk.model.MatchConfidence
+import com.emulnk.model.toWidgetConfig
 import com.google.gson.Gson
 import com.google.gson.reflect.TypeToken
-import java.util.concurrent.CancellationException
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.FlowPreview
 import kotlinx.coroutines.Job
@@ -53,7 +51,10 @@ import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import java.io.File
 import java.text.SimpleDateFormat
-import java.util.*
+import java.util.Date
+import java.util.Locale
+import java.util.UUID
+import java.util.concurrent.CancellationException
 
 class MainViewModel(application: Application) : AndroidViewModel(application) {
     companion object {
