+++ settings.gradle.kts
@@ -12,6 +12,10 @@ pluginManagement {
     }
 }
 
+plugins {
+    id("org.gradle.toolchains.foojay-resolver-convention") version "1.0.0"
+}
+
 dependencyResolutionManagement {
     repositoriesMode.set(RepositoriesMode.FAIL_ON_PROJECT_REPOS)
     repositories {
@@ -21,5 +25,4 @@ dependencyResolutionManagement {
 }
 
 rootProject.name = "EmuLink"
-include(":app", ":core")
- 
+include(":app", ":core", ":desktop")
